add bear_face/v3, folder maintenance * updated kb config.h * updated kb info.json, fixed version-level keyboard.json files * moved caps lock/backlight code to appropriate json files * updated SAFE_RANGE to QK_KB_0 * updated READMEs * added bear_face/v3 * removed old version C files (ex. v1.c etc.)
