+++ keyboards/bear_face/config.h
@@ -1,0 +1,4 @@
+/* Copyright 2024 will-hedges */
+/* SPDX-License-Identifier: GPL-2.0-or-later */
+
+#pragma once
