+++ keyboards/bear_face/v2/keymaps/default/keymap.c
@@ -1,19 +1,5 @@
-/*
-Copyright 2020 chemicalwill <https://github.com/chemicalwill>
-
-This program is free software: you can redistribute it and/or modify
-it under the terms of the GNU General Public License as published by
-the Free Software Foundation, either version 2 of the License, or
-(at your option) any later version.
-
-This program is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU General Public License for more details.
-
-You should have received a copy of the GNU General Public License
-along with this program.  If not, see <http://www.gnu.org/licenses/>.
-*/
+/* Copyright 2024 will-hedges */
+/* SPDX-License-Identifier: GPL-2.0-or-later */
 
 #include QMK_KEYBOARD_H
 
@@ -28,7 +14,7 @@ enum layers {
 
 //custom keycode enums
 enum custom_keycodes {
-    BASE_QWER = SAFE_RANGE,
+    BASE_QWER = QK_KB_0,
     BASE_COLE,
     BASE_DVOR
 };
@@ -82,8 +68,25 @@ const uint16_t PROGMEM keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
 */
 };
 
-//macros to allow the user to set whatever default layer they want, even after reboot
 
+// Caps Lock LED indicator
+void keyboard_pre_init_kb(void) {
+    //Sets LED pin as output
+    gpio_set_pin_output(F7);
+
+    keyboard_pre_init_user();
+};
+
+bool led_update_kb(led_t led_state) {
+    // Caps Lock LED indicator toggling code here
+    bool res = led_update_user(led_state);
+    if(res) {
+        gpio_write_pin(F7, led_state.caps_lock);
+    }
+    return res;
+};
+
+// macros to allow the user to set whatever default layer they want, even after reboot
 bool process_record_user(uint16_t keycode, keyrecord_t *record) {
     switch (keycode) {
         case BASE_QWER:
