+++ keyboards/bear_face/v3/keymaps/default_iso/readme.md
@@ -14,6 +14,5 @@ This layout replaces the stock layout on the Vortex Race 3.
     * 'APP' sends 'KC_APP'
     * Base layer toggles for QWERTY, COLEMAK, and DVORAK layouts (will persist after reboot)
 
-- New things in v2:
-    * Option of ANSI or ISO layout
-    * Optional stepped caps (will require compatible plate)
+- New things in v3:
+    * v3.1c PCB is compatible with both rev1 (Micro-USB) and rev2 (USB-C) stock cases
